COMP:Fixed warning about olf-style cast.

+++ VTK/Infovis/vtkBoxLayoutStrategy.cxx
@@ -80,7 +80,8 @@ void vtkBoxLayoutStrategy::Layout(vtkTree *inputTree,
     if (!inputTree->IsLeaf(vertex))
       {
       // Divide the available space with simple algo
-      int xDivisions = (int)(sqrt((float)nchildren)+1); // Ceiling
+      int xDivisions =
+        static_cast<int>(sqrt(static_cast<double>(nchildren))+1); // Ceiling
       int yDivisions = xDivisions;
        
       // Okay try shrinking the bounds
